// shesha-reactjs/src/apis/user.tsx
export interface ResetPasswordVerifyOtpResponse {
  /**
   * Indicates that the OTP matches to the sent one
   */
  isSuccess?: boolean;
  /**
   * Error message
   */
  errorMessage?: string | null;
  token?: string | null;
  username?: string | null;
}